fix: Integration with core
Done minor refactor of Auth to prevent from instantiating many times.
