+++ package.json
@@ -27,8 +27,10 @@
     "@ionic-native/splash-screen": "4.6.0",
     "@ionic-native/status-bar": "4.6.0",
     "@ionic/storage": "2.1.3",
+    "axios": "^0.18.0",
     "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
+    "lodash.find": "^4.6.0",
     "rxjs": "5.5.8",
     "sw-toolbox": "3.6.0",
     "zone.js": "0.8.26"
