+++ src/mobile-services.json
@@ -1,0 +1,35 @@
+{
+  "version": 1,
+  "clusterName": "192.168.64.74:8443",
+  "namespace": "myproject",
+  "clientId": "example_client_id",
+  "services": [
+    {
+      "id": "keycloak",
+      "name": "keycloak",
+      "type": "keycloak",
+      "url": "https://keycloak.security.feedhenry.org/auth/",
+      "config": {
+        "realm": "secure-app",
+        "url": "https://keycloak.security.feedhenry.org/auth/",
+        "ssl-required": "external",
+        "clientId": "client-app",
+        "public-client": true,
+        "use-resource-role-mappings": true,
+        "pinningFingerprint": "44 C8 9A 60 4E 29 82 85 8E 4F 75 1F 78 46 CD B3 0A 08 66 3F",
+        "apiServerUrl": "https://api.security.feedhenry.org",
+        "apiEndpoint": "/"
+      }
+    },
+    {
+      "id": "metrics",
+      "name": "metrics",
+      "type": "metrics",
+      "url": "https://www.mocky.io/v2/5aa696133100001335e716e0",
+      "config": {}
+    }
+  ]
+}
+
+
+
